refactor(frontend): currency item
